+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 All notable changes to this project will be documented in this file. This project adheres to Semantic Versioning.
 
+### 1.0.3
+* Function `getCmpResult` return only value of `CategoryMerchandisingResult`
+* Refactor logging and exception throwing 
+
 ### 1.0.2
 * Use `Product\Collection` instead of `Fulltext\Collection`
 * Add exceptions for easier debug